Use repeater element to make denser messages
