+++ packages/linear-genome-view/src/BasicTrack/components/ServerSideRenderedBlockContent.tsx
@@ -1,10 +1,11 @@
 import { makeStyles } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
-import { observer } from 'mobx-react'
+import { observer, PropTypes as MobxPropTypes } from 'mobx-react'
 import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
+import Button from '@material-ui/core/Button'
+import RefreshIcon from '@material-ui/icons/Refresh'
 import ServerSideRenderedContent from '../../LinearGenomeView/components/ServerSideRenderedContent'
-import BlockError from '../../LinearGenomeView/components/BlockError'
 
 const useStyles = makeStyles(theme => ({
   loading: {
@@ -24,11 +25,27 @@ const useStyles = makeStyles(theme => ({
     whiteSpace: 'normal',
   },
   blockMessage: {
+    width: '100%',
     background: theme.palette.action.disabledBackground,
     padding: theme.spacing(2),
     pointerEvents: 'none',
+    textAlign: 'center',
+  },
+  blockError: {
+    padding: theme.spacing(2),
+    pointerEvents: 'none',
+    width: '100%',
   },
 }))
+
+function Repeater({ children }: { children: React.ReactNode }) {
+  return (
+    <div style={{ display: 'flex' }}>
+      {children}
+      {children}
+    </div>
+  )
+}
 
 function LoadingMessage() {
   // only show the loading message after 300ms to prevent excessive flickering
@@ -64,6 +81,29 @@ BlockMessage.propTypes = {
   messageText: PropTypes.string.isRequired,
 }
 
+function BlockError({ error, reload }: { error: Error; reload: () => void }) {
+  const classes = useStyles()
+  return (
+    <div className={classes.blockError}>
+      {reload ? (
+        <Button onClick={reload} size="small" startIcon={<RefreshIcon />}>
+          Reload
+        </Button>
+      ) : null}
+      <Typography color="error" variant="body2">
+        {error.message}
+      </Typography>
+    </div>
+  )
+}
+BlockError.propTypes = {
+  error: MobxPropTypes.objectOrObservableObject.isRequired,
+  reload: PropTypes.func,
+}
+BlockError.defaultProps = {
+  reload: undefined,
+}
+
 const ServerSideRenderedBlockContent = observer(
   ({
     model,
@@ -77,13 +117,25 @@ const ServerSideRenderedBlockContent = observer(
     }
   }) => {
     if (model.error) {
-      return <BlockError error={model.error} reload={model.reload} />
+      return (
+        <Repeater>
+          <BlockError error={model.error} reload={model.reload} />
+        </Repeater>
+      )
     }
     if (model.message) {
-      return <BlockMessage messageText={model.message} />
+      return (
+        <Repeater>
+          <BlockMessage messageText={model.message} />
+        </Repeater>
+      )
     }
     if (!model.filled) {
-      return <LoadingMessage />
+      return (
+        <Repeater>
+          <LoadingMessage />
+        </Repeater>
+      )
     }
 
     return <ServerSideRenderedContent model={model} />
